requirements.txt
# ============================================================
# AI新闻自动化系统 — 依赖配置
# 推荐 Python 3.12.x
# ============================================================

# ----- Web框架 -----
Flask==3.1.0
Flask-CORS==5.0.1

# ----- 配置与环境 -----
PyYAML==6.0.2
python-dotenv==1.0.1

# ----- 浏览器自动化 -----
# Playwright 会自行解析兼容版本的 greenlet，
# 不要额外固定 greenlet 版本，否则会触发依赖冲突
playwright==1.49.1

# ----- HTTP请求与网页解析 -----
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
httpx==0.28.1

# ----- RSS解析 -----
feedparser==6.0.11

# ----- 数据处理 -----
pandas==2.2.3
SQLAlchemy==2.0.36

# ----- AI / LLM -----
openai==1.58.1

# ----- 图像处理 -----
Pillow==11.1.0

# ----- 任务调度 -----
APScheduler==3.11.0

# ----- 日志 -----
loguru==0.7.3

# ----- 微信公众号 -----
wechatpy==1.8.18
